requirements.txt

numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.5.0
scipy>=1.11.0

regex>=2023.10.3
spacy>=3.7.2
tqdm>=4.66.1



joblib>=1.4.0
matplotlib>=3.8.0
seaborn>=0.13.0


google-genai>=0.3.0


wandb>=0.23.0


python-dotenv>=1.0.0

setuptools>=70.0.0
wheel>=0.43.0
pip>=24.0
